requirements.txt
Flask>=3.0,<4.0
requests>=2.31,<3.0
pandas>=2.2,<3.0
networkx>=3.2,<4.0
python-dotenv>=1.0,<2.0
waitress>=3.0,<4.0
pytest>=8.0,<9.0
